paginate with button instead for testing
